+++ akshare/stock/stock_zh_a_tick_tx_163.py
@@ -43,7 +43,7 @@ def stock_zh_a_tick_tx_js(symbol: str = "sz000001") -> pd.DataFrame:
                 .str.split("/", expand=True)
             )
             page += 1
-            big_df_list.append(temp_df)
+            big_df = pd.concat([big_df, temp_df], ignore_index=True)
         except:
             break
     big_df = pd.concat(big_df_list, ignore_index=True)